add event tests
