+++ keeper/reconcile_test.go
@@ -1,6 +1,7 @@
 package keeper_test
 
 import (
+	"fmt"
 	"time"
 
 	sdkmath "cosmossdk.io/math"
@@ -11,12 +12,13 @@ import (
 )
 
 func (s *TestSuite) TestKeeper_ReconcileVaultInterest() {
+	twoMonths := -24 * 60 * time.Hour
 	shareDenom := "vaultshares"
 	underlying := sdk.NewInt64Coin("underlying", 1_000_000_000)
 	vaultAddress := types.GetVaultAddress(shareDenom)
 	testBlockTime := time.Now()
 	futureTime := testBlockTime.Add(100 * time.Second)
-	pastTime := testBlockTime.Add(-24 * 60 * time.Hour)
+	pastTime := testBlockTime.Add(twoMonths)
 
 	setup := func(interestRate string, periodStartSeconds int64) {
 		s.requireAddFinalizeAndActivateMarker(underlying, s.adminAddr)
@@ -41,6 +43,7 @@ func (s *TestSuite) TestKeeper_ReconcileVaultInterest() {
 				PeriodStart: periodStartSeconds,
 			}))
 		}
+		s.ctx = s.ctx.WithEventManager(sdk.NewEventManager())
 	}
 
 	tests := []struct {
@@ -48,6 +51,7 @@ func (s *TestSuite) TestKeeper_ReconcileVaultInterest() {
 		setup             func()
 		posthander        func()
 		expectedErrSubstr string
+		expectedEvents    sdk.Events
 	}{
 		{
 			name: "no start period found, should set period start and return no error",
@@ -58,6 +62,7 @@ func (s *TestSuite) TestKeeper_ReconcileVaultInterest() {
 				s.assertVaultInterestPeriod(vaultAddress, testBlockTime.Unix())
 				s.assertVaultAndMarkerBalances(vaultAddress, shareDenom, underlying.Denom, underlying.Amount, underlying.Amount)
 			},
+			expectedEvents: sdk.Events{},
 		},
 		{
 			name: "interest period start in future, should return nil and do nothing",
@@ -68,6 +73,7 @@ func (s *TestSuite) TestKeeper_ReconcileVaultInterest() {
 				s.assertVaultInterestPeriod(vaultAddress, futureTime.Unix())
 				s.assertVaultAndMarkerBalances(vaultAddress, shareDenom, underlying.Denom, underlying.Amount, underlying.Amount)
 			},
+			expectedEvents: sdk.Events{},
 		},
 		{
 			name: "interest period has elasped, should pay interest and update period start",
@@ -78,6 +84,7 @@ func (s *TestSuite) TestKeeper_ReconcileVaultInterest() {
 				s.assertVaultInterestPeriod(vaultAddress, testBlockTime.Unix())
 				s.assertVaultAndMarkerBalances(vaultAddress, shareDenom, underlying.Denom, sdkmath.NewInt(958047987), sdkmath.NewInt(1041952013))
 			},
+			expectedEvents: createReconcileEvents(vaultAddress, markertypes.MustGetMarkerAddress(shareDenom), sdkmath.NewInt(41952013), sdkmath.NewInt(1_000_000_000), sdkmath.NewInt(1041952013), underlying.Denom, "0.25", 5_184_000),
 		},
 		{
 			name: "interest period has elasped, should pay negative interest and update period start",
@@ -88,6 +95,7 @@ func (s *TestSuite) TestKeeper_ReconcileVaultInterest() {
 				s.assertVaultInterestPeriod(vaultAddress, testBlockTime.Unix())
 				s.assertVaultAndMarkerBalances(vaultAddress, shareDenom, underlying.Denom, sdkmath.NewInt(1040262904), sdkmath.NewInt(959737096))
 			},
+			expectedEvents: createReconcileEvents(vaultAddress, markertypes.MustGetMarkerAddress(shareDenom), sdkmath.NewInt(-40262904), sdkmath.NewInt(1_000_000_000), sdkmath.NewInt(959737096), underlying.Denom, "-0.25", 5_184_000),
 		},
 	}
 
@@ -109,24 +117,41 @@ func (s *TestSuite) TestKeeper_ReconcileVaultInterest() {
 				s.Require().Error(err)
 				s.Require().Contains(err.Error(), tc.expectedErrSubstr)
 			}
+
+			s.Assert().Equal(
+				normalizeEvents(tc.expectedEvents),
+				normalizeEvents(s.ctx.EventManager().Events()))
 		})
 	}
 }
 
-func (s *TestSuite) assertVaultInterestPeriod(vaultAddr sdk.AccAddress, expectedPeriod int64) {
-	interestDetails, err := s.k.VaultInterestDetails.Get(s.ctx, vaultAddr)
-	s.Require().NoError(err, "failed to get interest details")
-	s.Assert().Equal(expectedPeriod, interestDetails.PeriodStart, "unexpected interest period start")
-}
+func createReconcileEvents(vaultAddr, markerAddr sdk.AccAddress, interest, principle, principleAfter sdkmath.Int, denom, rate string, durations int64) []sdk.Event {
+	var allEvents []sdk.Event
 
-func (s *TestSuite) assertBalance(addr sdk.AccAddress, denom string, expectedAmt sdkmath.Int) {
-	balance := s.simApp.BankKeeper.GetBalance(s.ctx, addr, denom)
-	s.Assert().Equal(expectedAmt.String(), balance.Amount.String(), "unexpected balance for %s", addr.String())
-}
+	r, err := sdkmath.LegacyNewDecFromStr(rate)
+	if err != nil {
+		panic(fmt.Sprintf("invalid rate %s: %v", rate, err))
+	}
+	var fromAddress string
+	var toAddress string
+	if r.IsNegative() {
+		fromAddress = markerAddr.String()
+		toAddress = vaultAddr.String()
+	} else {
+		fromAddress = vaultAddr.String()
+		toAddress = markerAddr.String()
+	}
+	sendToMarkerEvents := createSendCoinEvents(fromAddress, toAddress, sdk.NewCoin(denom, interest.Abs()).String())
+	allEvents = append(allEvents, sendToMarkerEvents...)
 
-func (s *TestSuite) assertVaultAndMarkerBalances(vaultAddr sdk.AccAddress, shareDenom string, denom string, expectedVaultAmt, expectedMarkerAmt sdkmath.Int) {
-	markerAddr := markertypes.MustGetMarkerAddress(shareDenom)
-
-	s.assertBalance(vaultAddr, denom, expectedVaultAmt)
-	s.assertBalance(markerAddr, denom, expectedMarkerAmt)
+	reconcileEvent := sdk.NewEvent("vault.v1.EventVaultReconcile",
+		sdk.NewAttribute("interest_earned", CoinToJSON(sdk.Coin{Denom: denom, Amount: interest})),
+		sdk.NewAttribute("principal_after", CoinToJSON(sdk.NewCoin(denom, principleAfter))),
+		sdk.NewAttribute("principal_before", CoinToJSON(sdk.NewCoin(denom, principle))),
+		sdk.NewAttribute("rate", rate),
+		sdk.NewAttribute("time", fmt.Sprintf("%v", durations)),
+		sdk.NewAttribute("vault_address", vaultAddr.String()),
+	)
+	allEvents = append(allEvents, reconcileEvent)
+	return allEvents
 }
